tmux: set sensible environment defaults
fixes the tmux setup by addressing errors found in `:checkhealth tmux`.
the most notable of these errors is the inconsistent setup between the
expected `$TERM` variable and Tmux's internal default terminal. turns
out that we override the `TERM` variable in `~/.bash_profile`, which is
generally discouraged. the reason we have to do this though is because
of kitty's asinine `xterm-kitty` config. the terminfo is not portable to
other shells (like through SSH tunnels), and i really don't care enough
to fix this the "proper" way. i just want pretty colors in my terminals
:(