git(config): aliases for checkout + rebase
adds some aliases that allow the operator to checkout a branch and fast
forwards that branch to be up-to-date with remote ref.  takes care to
prevent destructive tendencies that might affect commit history. namely,
we only do ff merges and we do not use `git push` (whose behavior can be
dependent on git configs and provided flags).